+++ lib/rspec/core/formatters/deprecation_formatter.rb
@@ -4,30 +4,18 @@ module RSpec
   module Core
     module Formatters
       class DeprecationFormatter
-        attr_reader :count
+        attr_reader :count, :deprecation_stream, :summary_stream
 
-        def initialize(configuration)
-          @configuration = configuration
+        def initialize(deprecation_stream, summary_stream)
+          @deprecation_stream = deprecation_stream
+          @summary_stream = summary_stream
           @count = 0
         end
 
-        def deprecation_stream
-          @configuration.deprecation_stream || $stderr
-        end
-
-        def summary_stream
-          @configuration.output_stream || $stdout
-        end
-
         def printer
-          # to handle situations where we have preloaded early
-          if @configuration.output_stream
-            @printer ||= File === deprecation_stream ?
-              FilePrinter.new(deprecation_stream, summary_stream, self) :
-              IOPrinter.new(deprecation_stream, summary_stream, self)
-          else
+          @printer ||= File === deprecation_stream ?
+            FilePrinter.new(deprecation_stream, summary_stream, self) :
             IOPrinter.new(deprecation_stream, summary_stream, self)
-          end
         end
 
         def deprecation(data)