+++ spec/rspec/core/configuration_spec.rb
@@ -16,19 +16,6 @@ module RSpec::Core
       end
     end
 
-    describe '#output_stream' do
-      it 'is configurable' do
-        io = double 'output io'
-        config.output_stream = io
-        expect(config.output_stream).to eq io
-      end
-
-      it 'will load a file from a filename' do
-        config.output_stream = 'tmp/string.rb'
-        expect(config.output_stream).to be_a File
-      end
-    end
-
     describe '#deprecation_stream' do
       it 'defaults to standard error' do
         expect(config.deprecation_stream).to eq $stderr
@@ -39,10 +26,80 @@ module RSpec::Core
         config.deprecation_stream = io
         expect(config.deprecation_stream).to eq io
       end
+    end
 
-      it 'will load a file from a filename' do
-        config.deprecation_stream = 'tmp/string.rb'
-        expect(config.deprecation_stream).to be_a File
+    shared_examples_for "output_stream" do |attribute|
+      define_method :attribute_value do
+        config.__send__(attribute)
+      end
+
+      update_line = __LINE__ + 2
+      define_method :update_attribute do |value|
+        config.__send__(:"#{attribute}=", value)
+      end
+
+      it 'defaults to standard output' do
+        expect(attribute_value).to eq $stdout
+      end
+
+      it 'is configurable' do
+        io = double 'output io'
+        update_attribute(io)
+        expect(attribute_value).to eq io
+      end
+
+      context 'when the reporter has already been initialized' do
+        before do
+          config.reporter
+          allow(config).to receive(:warn)
+        end
+
+        it 'prints a notice indicating the reconfigured output_stream will be ignored' do
+          update_attribute(StringIO.new)
+          expect(config).to have_received(:warn).with(/output_stream.*#{__FILE__}:#{update_line}/)
+        end
+
+        it 'does not change the value of `output_stream`' do
+          update_attribute(StringIO.new)
+          expect(attribute_value).to eq($stdout)
+        end
+
+        it 'does not print a warning if set to the value it already has' do
+          update_attribute(attribute_value)
+          expect(config).not_to have_received(:warn)
+        end
+      end
+    end
+
+    describe "#output_stream" do
+      include_examples "output_stream", :output_stream
+    end
+
+    describe "#output" do
+      include_examples "output_stream", :output
+
+      specify 'the reader is deprecated' do
+        expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /output/)
+        config.output
+      end
+
+      specify 'the writer is deprecated' do
+        expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /output=/)
+        config.output = $stdout
+      end
+    end
+
+    describe "#out" do
+      include_examples "output_stream", :out
+
+      specify 'the reader is deprecated' do
+        expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /out/)
+        config.out
+      end
+
+      specify 'the writer is deprecated' do
+        expect_deprecation_with_call_site(__FILE__, __LINE__ + 1, /out=/)
+        config.out = $stdout
       end
     end
 
@@ -525,7 +582,7 @@ module RSpec::Core
             allow(Kernel).to receive(:warn)
             config.load_spec_files
             config.send(setter, "rspec/**/*.spec"); line = __LINE__
-            expect(Kernel).to have_received(:warn).with /has no effect.*#{__FILE__}:#{line}/
+            expect(Kernel).to have_received(:warn).with(/has no effect.*#{__FILE__}:#{line}/)
           end
 
           it 'will not warn if reset is called after load_spec_files' do
@@ -1185,14 +1242,6 @@ module RSpec::Core
       it "warns that this feature is deprecated" do
         expect_deprecation_with_call_site(__FILE__, __LINE__ + 1)
         config.debug?
-      end
-    end
-
-    describe "#output=" do
-      it "sets the output" do
-        output = double("output")
-        config.output = output
-        expect(config.output).to equal(output)
       end
     end